requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
plotly==5.22.0
scikit-learn==1.4.2
joblib==1.4.0
packaging==23.2  # Downgraded to compatible version
Pillow==10.3.0
streamlit-extras==0.3.6
pyarrow==15.0.2